ERTK-B1-smooth
ionosphere-fix: affected by the noise and ionospheric residual. (1,4,-5) and (0,-1,1) may be not suitable, one should try (1,0,-1) and (0,-1,1)
ionosphere-float: the position STDs of this model are lager than fix model. it is guessed that the second order of ionosphere is amplfied by the combination.
IN THIS UPDATE, add the B1 in float model, called smooth-B1. And the ambiguities information are delivered with reduced normal equation.

NOTE: the troposphere residual are not considered in those three model